Add -count-ifetch option to report the number of fetched instructions
The number of fetch instructions is equivalent to the number of
executed instructions if you consider that all fetched instructions
are executed even if they are invalidated, as it's typically the case
with predicated instructions.  This option is useful to evaluate some
compiler optimizations and is not related to the option "-icount".

Signed-off-by: Christophe Guillon <christophe.guillon@st.com>
Signed-off-by: Cédric VINCENT <cedric.vincent@st.com>