Add a new SH4 machine with one block of RAM and a "dummy" memory-mapped device.
This machine is mainly useful as an example on how to emulate a
memory-mapped device.  Currently the "dummy" memory-mapped device acts
almost like buffer: any non-NULL value written to an address can be
read back, NULL values are converted to random values.  Note that each
read/write event is printed in the standard output.

The RAM is located at the very beginning (0x00000000) and is
initialized with the content of the file "dummy.bin", the dummy device
is located at 0xffff0000.  The following SH4 program first reads 4
[random] values from the device, then write them back in the reverse
order:

    _start:
        mov.l   .SDI_address, r0
        mov.l   @r0+, r1
        mov.l   @r0+, r2
        mov.l   @r0+, r3
        mov.l   @r0+, r4
        mov.l   r1, @-r0
        mov.l   r2, @-r0
        mov.l   r3, @-r0
        mov.l   r4, @-r0

    .endless_loop:
        bra .endless_loop
        nop

        .align 2
    .dummy_address:
        .long   0xffff0000

You can build it that way:

    $ sh4-linux-as dummy.s -o dummy.elf
    $ sh4-linux-objcopy -O binary dummy.elf dummy.bin

It produces these events:

    $ qemu-system-sh4 -M dummy
    @0 -> 1804289383 (4)
    @4 -> 846930886 (4)
    @8 -> 1681692777 (4)
    @12 -> 1714636915 (4)
    @12 <- 1804289383 (4)
    @8 <- 846930886 (4)
    @4 <- 1681692777 (4)
    @0 <- 1714636915 (4)

To use this machine with another CPU, you may only have to adjust the
name of the program-counter register [1], "pc" in this line:

    cpu_state->pc = reset_address;

[1]: assuming the MMU support is already there.