linux-user: enable forward compatibility for host syscalls.
Added an --enabled-syscall-fwd-compat option in order
to enable forward compatibility of syscalls in linux-user mode.

It means for instance that if QEMU is built on a host
with a old kernel and this option is enabled
it will execute syscalls when running on a more
recent host.
If the syscall is actually not supported on the actual
host running QEMU, the behavior is undefined, generally
the syscall will fail.
The standard behavior without this option
is that QEMU skips the syscall ("unsupported syscall").

A typical example case is the support for *at syscalls
introduced in kernel 2.6.16. For instance if QEMU is
compiled on a CentOS4 (kernel 2.6.9), the same binary
will be also usable on a CentOS5 (kernel 2.6.18) running
a guest system with a 2.6.16+ kernel.
Actually when executing an openat syscall on the 2.6.16+
guest, QEMU will forward it to the openat syscall on the
CenOS5 kernel.
Note that it is different from syscall emulation, for
instance running the same 2.6.16+ guest system on the
initial CentOS4 2.6.9 kernel will fail as the openat
syscall is actually not supported.

In order to activate forward compatibility, we just have
to define when it is not done by the build host the syscall
numbers as extracted from the target syscall number files.

Change-Id: If839a26833f4d9cfcbcb41651514ef71b0422855
Signed-off-by: Christophe Guillon <christophe.guillon@st.com>