sh4-linux-user: Add partial support for gUSA ABI mechanism.
As of my understanding of:

    http://www.free-it.org/archiv/LT_History/2003/talks/170/paper.html

When a thread with a negative stack pointer (R15) is preempted
(because of a signal or a reschedule), the Linux/SH kernel has to
restart it from the address R0 + R15 - 2.  This "restartable code
sequence" mechanism is the way atomic operations are handled on
Linux/SH and is heavily used by the user-space part of futex support.

When the thread was preempted because of signal, the kernel has to
restore the original value of the stack pointer from R1 before
jumping to the signal handler.  Signal handlers are not part of the
critical section and requires a valid stack pointer.

This gUSA mechanism works only on non-preemptible kernels running on
top of mono-processors.  Sadly, this is obviously not the case when
using QEMU user-mode on modern multi-core workstation, as a
consequence we have to use a locking mechanism:

    lock(gUSA_lock)
    ... gUSA prologue
    ... restartable code sequence
    ... gUSA epilogue
    unlock(gUSA_lock)

Not done yet:

    * restore the stack pointer when a guest signal handler is called,
      otherwise it might crash when running in single step or debug
      mode. (in nominal mode, the call to guest signal handlers are
      delayed at the end of a translated block).

    * use atomic writes on the host side, this is required by the gUSA
      ABI mechanism.

    * it doesn't work on memory shared by two process (i.e. not
      thread) yet. This is the case for most atomic memory operation
      on QEMU.  (a solution could be the use of PRoot, since it
      monitores all the processess).

Note: QEMU will dead-lock if this mechanism is used in a signal
handler, however we do known that signals and threads don't mix well
on QEMU user-mode.

Change-Id: I1f1e78ca5f2611737941ee01fb061cc95d9de119