sh4-linux-user: Don't try to emulate the MOVCAL instruction accurately.
The current emulation of the MOVCAL instruction is broken on
multi-threaded programs:

    *** glibc detected *** qemu-sh4: malloc(): double linked list corrupted: 0x0000000062227160 ***
    ======= Backtrace: =========
    /lib/libc.so.6(+0x775b6)[0x7ffff6e375b6]
    /lib/libc.so.6(+0x7b8e9)[0x7ffff6e3b8e9]
    /lib/libc.so.6(__libc_malloc+0x6e)[0x7ffff6e3c58e]
    qemu-sh4(helper_movcal+0x2a)[0x6004e0ba]
    qemu-sh4[0x6021d233]

Moreover the accurate emulation of this instruction is useless in
user-mode: it moves a word in the cache without refilling the memory
line at the given address, then the cached line is written back to the
memory according to the policy eviction unless an OCBI instruction is
performed before.  As a consequence this instruction is useful only in
kernel code to implement the cache flush mechanism (the cache behavior
is unpredictable in user-space).

Change-Id: I380d2be436bf82ba285424c955b55c20b2998dd3