TCG: increase opcode/param buffer size when plugins are enabled.
When CONFIG_ENABLE_PLUGIN is defined, each initial TCG operation
can be instrumented by a plugin. Hence the static setting for
the maximum number of opcodes/parameters can be overflowed.
Defined a fixed additional plugin operation count per initial
operation and account for it in the opcodes/parameters buffers
size computation when this configuration is active.

Change-Id: Ia060005e6df40c74c39da4a314120bdf44cc69a6